requirements.txt
numpy
tensorflow
scikit-learn
datasets
pillow
datetime
# torch
# huggingface_hub
# h5py